+++ packages/axios-extension/src/abap/index.ts
@@ -11,4 +11,4 @@ export { AppIndex, AppIndexService } from './app-index-service';
 export * from './message';
 export * from './catalog';
 export * from './adt-catalog';
-export { TenantType } from './types';
+export { TenantType, ArchiveFileNode } from './types';