+++ packages/ui5-application-writer/src/types.ts
@@ -73,6 +73,11 @@ export interface AppOptions {
      * Enable support for importing functionality from npm packages in UI5 applications
      */
     npmPackageConsumption: boolean;
+
+    /**
+     * Excludes the index.html from the template and does not add the `start-noflp` script in package.json
+     */
+    generateIndex?: boolean;
 }
 
 export interface Ui5App {