+++ packages/ui5-application-writer/templates/optional/loadReuseLibs/webapp/test/locate-reuse-libs.js
@@ -65,7 +65,7 @@
                     }
                     resolve(result);
                 })
-                .fail(function (error) {
+                .fail(function () {
                     reject(new Error("Could not fetch manifest at '" + manifestPath));
                 });
         });
@@ -131,7 +131,7 @@
     };
 })(sap);
 
-/*eslint-disable sap-browser-api-warning, sap-no-dom-access*/
+/*eslint-disable fiori-custom/sap-browser-api-warning, fiori-custom/sap-no-dom-access*/
 var currentScript = document.getElementById("locate-reuse-libs");
 if (!currentScript) {
     currentScript = document.currentScript;