+++ packages/axios-extension/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @sap-ux/axios-extension
 
+## 1.2.2
+
+### Patch Changes
+
+-   fa94bfd6: Only eject the fetch request interceptor when a valid csrf token is received
+
+## 1.2.1
+
+### Patch Changes
+
+-   3d3d8c64: Fixes for unsage usage of optional chaining sonar bugs
+
+## 1.2.0
+
+### Minor Changes
+
+-   c775d787: This change implements a new ADT service `FileStoreService`.
+    `FileStoreService` supports querying the file structure and file content in a deployed Fiori app archive.
+
+    Example use case:
+
+    ```
+    const fileStoreService = await provider.getAdtService<FileStoreService>(FileStoreService);
+    // Fetch a list of files and folders in the app's root folder.
+    const rootFolderContent = await fileStoreService.getAppArchiveContent('folder' 'ZFIORIAPP');
+    // Fetch a list of files and folders in <root>/webapp
+    const webappFolderContent = await fileStoreService.getAppArchiveContent('folder' 'ZFIORIAPP', '/webapp');
+    // Fetch the text content as string from <root>/package.json file.
+    const fileContent = await fileStoreService.getAppArchiveContent('file' 'ZFIORIAPP', '/package.json');
+    ```
+
 ## 1.1.0
 
 ### Minor Changes