+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.18.9
+
+### Patch Changes
+
+-   1522b416: Remove redundant commas from templates
+
+## 0.18.8
+
+### Patch Changes
+
+-   3391be7f: Feature enhancement: in addition to the entity (set) you now can define a contextPath that gets written instead of entitySet into the `options > settings` of the new page.
+
 ## 0.18.7
 
 ### Patch Changes